requirements.txt
playwright>=1.40.0
supabase>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
cloudscraper>=1.2.71
